requirements.txt
Flask==2.0.3
flask-cors==3.0.10
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
Werkzeug==2.0.3
gunicorn==20.1.0 